requirements.txt
aiohttp
matrix-nio
ollama
requests
rich
